# internal/pkg/models/book.go
/*
 * Brock's Library
 *
 * REST API for Managing Library Books
 *
 * API version: 0.1.0
 * Contact: brock@email.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */
package models

import (
	"time"
)

type Book struct {

	// The object ID of resource
	Id string `json:"id"`

	// the official title of the book
	Title string `json:"title"`

	// A list of authors who wrote the book
	Authors []string `json:"authors"`

	// the year the book was originally published
	Year int32 `json:"year"`

	// The identifier for a specific version of the resource
	Version string `json:"version,omitempty"`

	// The UTC time the object was created
	CreationTime time.Time `json:"creationTime,omitempty"`

	// The UTC time the object was updated
	LastUpdateTime time.Time `json:"lastUpdateTime,omitempty"`

	// Comments about the book
	Comment string `json:"comment,omitempty"`
}