# internal/pkg/models/error.go
/*
 * Brock's Library
 *
 * REST API for Managing Library Books
 *
 * API version: 0.1.0
 * Contact: brock@email.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type Error struct {
	Code int32 `json:"code"`

	Message string `json:"message"`
}